+++ source/js/common/columns-dialog.js
@@ -8,6 +8,7 @@ let initialized = false;
  * @typedef ColumnDescription
  * @type {Object}
  * @property {string} name - column name
+ * @property {string} customElement - synthetic custom column name for saving visibility in localStorage
  * @property {string} element - property name template (could be ends with "[x]") of resource object
  * @property {string} types - possible value types in the column
  * @property {boolean} isArray - true if max cardinality greater than 1