+++ source/js/reporter.js
@@ -186,11 +186,18 @@ ${measurement.name}:
   }
 
   /**
+   * @typedef MeasurementController
+   * Object to control the measurement.
+   * @property {function(number): void} updateCount - update measurement value
+   * @property {function(number=): void} incrementCount - increase measurement value
+   * @property {function(): number} getCount - get measurement value
+   */
+  /**
    * Records the start of metric measurement. Returns an object to control the measurement.
    * @param {string} name
    * @param {boolean} calculateDuration
    * @param {number} count
-   * @return {{updateCount: (function(number): void), incrementCount: (function(number): void)}}
+   * @return {MeasurementController}
    */
   addMetric({ name, calculateDuration = true, count = 0 }) {
     const measurement = {
@@ -219,10 +226,9 @@ ${measurement.name}:
       updateCount,
       /**
        * Increment current measurement count
-       * @param {number} inc
+       * @param {number} [inc]
        */
-      incrementCount: (inc) =>
-        updateCount(measurement.count + (typeof inc === 'number' ? inc : 1)),
+      incrementCount: (inc = 1) => updateCount(measurement.count + inc),
       /**
        * Returns current measurement count
        * @return {number}